refactor: Move source maps out of bundle
The newly added encode map remains, everything else is removed.
