+++ .prettierignore
@@ -2,3 +2,4 @@ node_modules/
 coverage/
 lib/
 src/maps/
+maps/
